add permission to hall

+++ halls/models.py
@@ -44,7 +44,7 @@ class Hall(models.Model):
     moderated = models.BooleanField(default=False)
     owner = models.ForeignKey(User, on_delete=models.SET_NULL, related_name='halls', null=True, blank=True)
     hall_type = models.ManyToManyField(HallType, related_name='halls', blank=True)
-    view_count = models.IntegerField(default=0)
+    view_count = models.BigIntegerField(default=0)
     area = models.DecimalField(max_digits=100, decimal_places=2, null=True,)
     capacity = models.IntegerField(null=True)
     rating = models.DecimalField(max_digits=3, decimal_places=2, null=True, default=0,)
